+++ examples/segmentation_3d/unet_evaluation_dict.py
@@ -23,7 +23,7 @@ from torch.utils.data import DataLoader
 import monai
 from monai import config
 from monai.data.utils import list_data_collate
-from monai.utils.sliding_window_inference import sliding_window_inference
+from monai.data.sliding_window_inference import sliding_window_inference
 from monai.metrics.meandice import compute_meandice
 from monai.data.synthetic import create_test_image_3d
 from monai.networks.nets.unet import UNet
