+++ examples/segmentation_3d/unet_training_array.py
@@ -26,7 +26,7 @@ import monai.transforms.compose as transforms
 from monai.data.nifti_reader import NiftiDataset
 from monai.transforms import AddChannel, Rescale, RandUniformPatch, RandRotate90, ToTensor
 from monai.data.synthetic import create_test_image_3d
-from monai.utils.sliding_window_inference import sliding_window_inference
+from monai.data.sliding_window_inference import sliding_window_inference
 from monai.metrics.meandice import compute_meandice
 from monai.visualize.img2tensorboard import plot_2d_or_3d_image
