+++ monai/data/__init__.py
@@ -15,3 +15,6 @@ from .grid_dataset import GridPatchDataset
 from .nifti_reader import load_nifti, NiftiDataset
 from .nifti_saver import NiftiSaver
 from .nifti_writer import write_nifti
+from .sliding_window_inference import sliding_window_inference
+from .synthetic import *
+from .utils import *
