+++ monai/handlers/__init__.py
@@ -16,3 +16,4 @@ from .metric_logger import *
 from .segmentation_saver import SegmentationSaver
 from .stats_handler import StatsHandler
 from .tensorboard_handlers import TensorBoardImageHandler, TensorBoardStatsHandler
+from .utils import *
