+++ monai/networks/blocks/__init__.py
@@ -8,3 +8,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from .convolutions import *
