+++ monai/networks/layers/__init__.py
@@ -8,3 +8,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from .convutils import *
+from .factories import *
+from .simplelayers import *
