+++ monai/transforms/utils.py
@@ -39,12 +39,6 @@ def is_empty(img):
     return not (img.max() > img.min())  # use > instead of <= so that an image full of NaNs will result in True
 
 
-def ensure_tuple_size(tup, dim):
-    """Returns a copy of `tup` with `dim` values by either shortened or padded with zeros as necessary."""
-    tup = tuple(tup) + (0,) * dim
-    return tup[:dim]
-
-
 def zero_margins(img, margin):
     """Returns True if the values within `margin` indices of the edges of `img` in dimensions 1 and 2 are 0."""
     if np.any(img[:, :, :margin]) or np.any(img[:, :, -margin:]):
@@ -364,38 +358,3 @@ def create_translate(spatial_dims, shift):
     return affine
 
 
-def to_affine_nd(r, affine):
-    """
-    Using elements from affine, to create a new affine matrix by
-    assigning the rotation/zoom/scaling matrix and the translation vector.
-
-    when ``r`` is an integer, output is an (r+1)x(r+1) matrix,
-    where the top left kxk elements are copied from ``affine``,
-    the last column of the output affine is copied from ``affine``'s last column.
-    `k` is determined by `min(r, len(affine) - 1)`.
-
-    when ``r`` is an affine matrix, the output has the same as ``r``,
-    the top left kxk elments are  copied from ``affine``,
-    the last column of the output affine is copied from ``affine``'s last column.
-    `k` is determined by `min(len(r) - 1, len(affine) - 1)`.
-
-
-    Args:
-        r (int or matrix): number of spatial dimensions or an output affine to be filled.
-        affine (matrix): 2D affine matrix
-    Returns:
-        a (r+1) x (r+1) matrix
-    """
-    affine = np.array(affine, dtype=np.float64)
-    if affine.ndim != 2:
-        raise ValueError('input affine must have two dimensions')
-    new_affine = np.array(r, dtype=np.float64, copy=True)
-    if new_affine.ndim == 0:
-        sr = new_affine.astype(int)
-        if not np.isfinite(sr) or sr < 0:
-            raise ValueError('r must be postive.')
-        new_affine = np.eye(sr + 1, dtype=np.float64)
-    d = min(len(new_affine) - 1, len(affine) - 1)
-    new_affine[:d, :d] = affine[:d, :d]
-    new_affine[:d, -1] = affine[:d, -1]
-    return new_affine
