+++ tests/test_integration_determinism.py
@@ -15,11 +15,10 @@ import numpy as np
 import torch
 from torch.utils.data import DataLoader, Dataset
 
-import monai.transforms.compose as transforms
-from monai.data.synthetic import create_test_image_2d
-from monai.losses.dice import DiceLoss
-from monai.networks.nets.unet import UNet
-from monai.transforms import AddChannel, RandRotate90, RandUniformPatch, Rescale, ToTensor
+from monai.data import create_test_image_2d
+from monai.losses import DiceLoss
+from monai.networks.nets import UNet
+from monai.transforms import Compose, AddChannel, RandRotate90, RandUniformPatch, Rescale, ToTensor
 
 
 def run_test(batch_size=64, train_steps=200, device=torch.device("cuda:0")):
@@ -52,8 +51,7 @@ def run_test(batch_size=64, train_steps=200, device=torch.device("cuda:0")):
 
     loss = DiceLoss(do_sigmoid=True)
     opt = torch.optim.Adam(net.parameters(), 1e-2)
-    train_transforms = transforms.Compose([AddChannel(), Rescale(), RandUniformPatch((96, 96)),
-                                           RandRotate90(), ToTensor()])
+    train_transforms = Compose([AddChannel(), Rescale(), RandUniformPatch((96, 96)), RandRotate90(), ToTensor()])
 
     src = DataLoader(_TestBatch(train_transforms), batch_size=batch_size)
