+++ tests/test_sliding_window_inference.py
@@ -14,7 +14,7 @@ import numpy as np
 import torch
 from parameterized import parameterized
 
-from monai.utils.sliding_window_inference import sliding_window_inference
+from monai.data.sliding_window_inference import sliding_window_inference
 
 TEST_CASE_1 = [(1, 3, 16, 15, 7), (4, 10, 7), 3]  # 3D small roi
